# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Pangolin Catalog API
  description: Multi-tenant Apache Iceberg REST Catalog with advanced features including federated catalogs, credential vending, and fine-grained permissions
  contact:
    name: Pangolin API Support
    email: support@pangolin.dev
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
paths:
...
  /api/v1/service-users:
    get:
      tags:
      - Service Users
      summary: List service users for the tenant
      description: GET /api/v1/service-users
      operationId: list_service_users
      responses:
        '200':
          description: List of service users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ServiceUser'
        '403':
          description: Forbidden
        '500':
          description: Internal server error
      security:
      - bearer_auth: []
    post:
      tags:
      - Service Users
      summary: Create a new service user
      description: POST /api/v1/service-users
      operationId: create_service_user
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateServiceUserRequest'
        required: true
      responses:
        '201':
          description: Service user created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyResponse'
        '403':
          description: Forbidden
        '500':
          description: Internal server error
      security:
      - bearer_auth: []
  /api/v1/service-users/{id}:
    get:
      tags:
      - Service Users
      summary: Get a specific service user
      description: GET /api/v1/service-users/{id}
      operationId: get_service_user
      parameters:
      - name: id
        in: path
        description: Service user ID
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Service user details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceUser'
        '403':
          description: Forbidden
        '404':
          description: Service user not found
        '500':
          description: Internal server error
      security:
      - bearer_auth: []
    put:
      tags:
      - Service Users
      summary: Update a service user
      description: PUT /api/v1/service-users/{id}
      operationId: update_service_user
      parameters:
      - name: id
        in: path
        description: Service user ID
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateServiceUserRequest'
        required: true
      responses:
        '200':
          description: Service user updated
        '403':
          description: Forbidden
        '500':
          description: Internal server error
      security:
      - bearer_auth: []
    delete:
      tags:
      - Service Users
      summary: Delete a service user
      description: DELETE /api/v1/service-users/{id}
      operationId: delete_service_user
      parameters:
      - name: id
        in: path
        description: Service user ID
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Service user deleted
        '403':
          description: Forbidden
        '500':
          description: Internal server error
      security:
      - bearer_auth: []
  /api/v1/service-users/{id}/rotate:
    post:
      tags:
      - Service Users
      summary: Rotate API key for a service user
      description: POST /api/v1/service-users/{id}/rotate
      operationId: rotate_api_key
      parameters:
      - name: id
        in: path
        description: Service user ID
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: API key rotated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyResponse'
        '403':
          description: Forbidden
        '404':
          description: Service user not found
        '500':
          description: Internal server error
      security:
      - bearer_auth: []
...
components:
  schemas:
...
    ApiKeyResponse:
      type: object
      description: API Key response (only shown once on creation)
      required:
      - service_user_id
      - name
      - api_key
      properties:
        api_key:
          type: string
        expires_at:
          type: string
          format: date-time
          nullable: true
        name:
          type: string
        service_user_id:
          type: string
          format: uuid
...
    CreateServiceUserRequest:
      type: object
      required:
      - name
      - role
      properties:
        description:
          type: string
          nullable: true
        expires_in_days:
          type: integer
          format: int64
          nullable: true
        name:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
...
    ServiceUser:
      type: object
      description: Service user for API key authentication
      required:
      - id
      - name
      - tenant-id
      - api-key-hash
      - role
      - created-at
      - created-by
      - active
      properties:
        active:
          type: boolean
        api-key-hash:
          type: string
        created-at:
          type: string
          format: date-time
        created-by:
          type: string
          format: uuid
        description:
          type: string
          nullable: true
        expires-at:
          type: string
          format: date-time
          nullable: true
        id:
          type: string
          format: uuid
        last-used:
          type: string
          format: date-time
          nullable: true
        name:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        tenant-id:
          type: string
          format: uuid
...
    UpdateServiceUserRequest:
      type: object
      properties:
        active:
          type: boolean
          nullable: true
        description:
          type: string
          nullable: true
        name:
          type: string
          nullable: true
...
    UserRole:
      type: string
      description: User role in the system
      enum:
      - root
      - tenant-admin
      - tenant-user
...
  securitySchemes:
    bearer_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/v1/users/login or /api/v1/tokens
...